# test: fix file system race conditions in tests by switching to an IPC server