--- a/__utils__/test-ipc-server/bin/test-ipc-server-client.js
+++ b/__utils__/test-ipc-server/bin/test-ipc-server-client.js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+require('../lib/clientBin')