--- a/__utils__/test-ipc-server/jest.config.js
+++ b/__utils__/test-ipc-server/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require('../../jest.config')