--- a/__utils__/test-ipc-server/src/clientBin.ts
+++ b/__utils__/test-ipc-server/src/clientBin.ts
@@ -1,0 +1,9 @@
+import net from 'node:net'
+import { computeHandlePath } from './computeHandlePath'
+
+const [handle] = process.argv.slice(2)
+const connectPath = computeHandlePath(handle)
+
+const client = net.connect(connectPath, () => {
+  process.stdin.pipe(client).on('end', () => client.destroy())
+})