--- a/__utils__/test-ipc-server/src/index.ts
+++ b/__utils__/test-ipc-server/src/index.ts
@@ -1,0 +1,1 @@
+export { createTestIpcServer, TestIpcServer } from './TestIpcServer'