--- a/exec/lifecycle/test/fixtures/.gitignore
+++ b/exec/lifecycle/test/fixtures/.gitignore
@@ -1,1 +1,2 @@
 output.json
+test.sock