--- a/pkg-manager/headless/test/fixtures/deps-have-lifecycle-scripts/.gitignore
+++ b/pkg-manager/headless/test/fixtures/deps-have-lifecycle-scripts/.gitignore
@@ -1,1 +1,1 @@
-output.json
+test.sock