… __utils__/test-ipc-server/README.md
… __utils__/test-ipc-server/package.json
… __utils__/test-ipc-server/tsconfig.json
… __utils__/test-ipc-server/tsconfig.lint.json
… exec/lifecycle/package.json
… exec/lifecycle/test/fixtures/gyp-with-preinstall/package.json
… exec/lifecycle/test/fixtures/inspect-frozen-lockfile/package.json
… exec/lifecycle/test/fixtures/simple/package.json
… exec/lifecycle/test/fixtures/with-many-scripts/package.json
… exec/lifecycle/tsconfig.json
… exec/plugin-commands-rebuild/package.json
… exec/plugin-commands-rebuild/tsconfig.json
… exec/plugin-commands-script-runners/package.json
… exec/plugin-commands-script-runners/test/testRecursive.ts
… exec/plugin-commands-script-runners/tsconfig.json
… exec/prepare-package/package.json
… exec/prepare-package/test/__fixtures__/has-prepublish-script-and-main-file/package.json
… exec/prepare-package/test/__fixtures__/has-prepublish-script/package.json
… exec/prepare-package/tsconfig.json
… pkg-manager/core/package.json
… pkg-manager/core/tsconfig.json
… pkg-manager/headless/package.json
… pkg-manager/headless/test/fixtures/deps-have-lifecycle-scripts/package.json
… pkg-manager/headless/test/fixtures/deps-have-lifecycle-scripts/pnpm-lock.yaml
… pkg-manager/headless/test/index.ts
… pkg-manager/headless/tsconfig.json
… pkg-manager/plugin-commands-installation/package.json
… pkg-manager/plugin-commands-installation/test/dedupe.ts
… pkg-manager/plugin-commands-installation/tsconfig.json
… pnpm-lock.yaml